refactor: rename manager to PathWatcher
