+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/ClientController.scala
@@ -14,7 +14,7 @@ import org.enso.languageserver.capability.CapabilityApi.{
 import org.enso.languageserver.capability.CapabilityProtocol
 import org.enso.languageserver.data.Client
 import org.enso.languageserver.event.{ClientConnected, ClientDisconnected}
-import org.enso.languageserver.filemanager.FileEventManagerProtocol
+import org.enso.languageserver.filemanager.PathWatcherProtocol
 import org.enso.languageserver.filemanager.FileManagerApi._
 import org.enso.languageserver.requesthandler._
 import org.enso.languageserver.text.TextApi._
@@ -97,7 +97,7 @@ class ClientController(
     case TextProtocol.TextDidChange(changes) =>
       webActor ! Notification(TextDidChange, TextDidChange.Params(changes))
 
-    case FileEventManagerProtocol.FileEventResult(event) =>
+    case PathWatcherProtocol.FileEventResult(event) =>
       webActor ! Notification(EventFile, EventFile.Params(event))
 
     case r @ Request(method, _, _) if (requestHandlers.contains(method)) =>
